requirements.txt
asgiref==3.11.1
Django==6.0.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.10.2
pillow==12.1.0
PyJWT==2.11.0
sqlparse==0.5.5
tzdata==2025.3